backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Vector Database & Embeddings
qdrant-client==1.7.3
sentence-transformers==2.3.1

# LLM Integration
anthropic==0.26.0
groq
# Document Processing
pypdf==4.0.1
python-docx==1.1.0
markdown==3.5.2

# Environment & Configuration
python-dotenv==1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# CORS & Security
python-jose[cryptography]==3.3.0

# Utilities
aiofiles==23.2.1